requirements.txt
Cython==3.0.8
future==0.18.3
h5py==3.10.0
matplotlib==3.7.5
numpy==1.24.4
pillow==10.2.0
PyQt6==6.4.2
pyserial==3.5
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.10.1
serial==0.0.97
setuptools~=65.5.1